derp (forgot git add)

+++ Blackboard/Blackboard.cpp
@@ -263,7 +263,7 @@ void Blackboard::handlePacket(int fd, const Packet & packet) {
             for (std::deque<DataPoint>::const_iterator it = retSet.begin(); it != retSet.end(); ++it) {
                 pktSize += it->size() + sizeof (uint32_t); // add datapoint size + size of datapoint size;
             }
-            log("pktSize = %d. retSet.size() = %lu.\n", pktSize, retSet.size());
+            log("pktSize = %d. retSet.size() = %d.\n", pktSize, (int)retSet.size());
 
             if (pktSize > MAX_BUFFER_SIZE) {
                 log("Requested data set is too large! (%d)\n", (int) retSet.size());
